Added Delete functionality

+++ resources/views/livewire/notes/show.blade.php
@@ -1,12 +1,16 @@
 <?php
 
 use Livewire\Volt\Component;
+use App\Models\Note;
 
 new class extends Component {
+
     public function delete($noteId) : void {
         $note = Note::where('id', $noteId)->first();
+        $this->authorize('delete', $note);
         $note->delete();
     }
+    
     public function with(): array
     {
         return [
@@ -31,20 +35,24 @@ new class extends Component {
             <div class="grid grid-cols-3 gap-4 mt-12">
                 @foreach ($notes as $note)
                     <x-card wire:key='{{$note -> id}}' >
-                        <div class="flex justify-between">
-                            <div>
-                                <a href="#" class="text-xl font-bold hover:underline hover:text-blue-500">{{$note->title}}</a>
-                                <p class="mt-2 text-xs">{{Str::limit($note->body, 100)}}</p>
+                        <div class="flex justify-between gap-2">
+                            <div >
+                                <a href="{{route('notes.edit', $note)}}" wire:navigate class="text-xl font-bold hover:underline hover:text-blue-500">
+                                    {{$note->title}}
+                                </a>
+                                <p class="mt-2 text-xs">{{Str::limit($note->body, 50)}}</p>
                             </div>
-                            <div class="text-xs text-gray-500 ">{{\Carbon\Carbon::parse($note->send_date)->format("d-M-Y")}}</div>
+                            <div class="text-xs text-gray-500 text-pretty lg:text-nowrap lg:w-full w-">
+                                {{\Carbon\Carbon::parse($note->send_date)->format("d-M-Y")}}
+                            </div>
                         </div>
                         <div class="flex items-end justify-between mt-4 space-y-1">
                             <p class="text-xs">Recipient: <span class="font-semibold"> {{$note->recipient}}
                                 </span>
                             </p>
                             <div>
-                                <x-button.circle icon="eye"></x-button>
-                                <x-button.circle icon="trash"></x-button>
+                                <x-button.circle icon="eye" ></x-button>
+                                <x-button.circle icon="trash" wire:click="delete('{{$note->id}}')"></x-button>
                             </div>
                         </div>
                     </x-card>
